[Feat] 파사드 패턴을 통해 순환참조 해결
